refactor(T11527): delete legacy data-safety.ts → data-safety-central.ts; remove repairMissingColumns (E6-L7)
Merges the task-level safety wrappers (SafetyConfig/SafetyError, safeCreateTask/
safeUpdateTask/safeDeleteTask, checkTaskExists, verifyTaskWrite, verifySessionWrite,
validateAndRepairSequence, triggerCheckpoint, safeCreateSession,
forceCheckpointBeforeOperation) from the former data-safety.ts into
data-safety-central.ts (single point of safety). The cwd-only integrity probe is
renamed runSequenceIntegrityCheck to avoid colliding with the accessor-based
runDataIntegrityCheck. Reroutes tasks-sqlite.ts + store/index.ts + 5 safety tests;
deletes data-safety.ts (zero importers).

Removes repair.ts::repairMissingColumns and drops it from runAllRepairs — DDL is
now owned exclusively by the forward Drizzle migrations, not post-hoc ensureColumns.

Validation: pnpm run typecheck clean; build + CLI smoke pass; @cleocode/core 15
pre-existing worktree/napi flakes (deterministic ×2, none touch changed code),
zero NEW; @cleocode/runtime 120/120; cleo check arch 5/5 (DB Open Guard green).
